Improve resource usage by filtering out pulls that won't be displayed
There was logic in the CSS file to set `display: none` for PRs marked with the `unimportant-repo` and `unimportant-user` classes.
This was inefficient as only a small number of all PRs would end up being displayed if user filtering was used. This caused high resource usage in the browser as it would spend time rendering and keeping in memory a lot of views that would never be displayed.

It also caused a lot of unnecessary calls to the GitHub API and could result in violating the rate limit.

This change filters the PRs after they are fetched and before comments and additional info are fetched, which reduces the number of API calls. Only that small group of PRs will be shown and they are the only ones than get rendered, which saves browser memory and resource usage.

Added tests to check PR filtering.

Author: @CharlesIC